+++ soloist-player-service/src/main/kotlin/com/soloist/player/kafka/consumer/SaveTasksConsumer.kt
@@ -11,7 +11,7 @@ import com.soloist.player.model.entity.Immutables
 import com.soloist.player.model.entity.task.Task
 import com.soloist.player.model.entity.task.dto.SaveTaskInput
 import com.soloist.player.model.entity.task.enums.TaskTopic
-import com.soloist.player.service.ai.TaskVectorService
+import com.soloist.player.service.task.impl.TaskVectorService
 import com.soloist.player.service.task.PlayerTaskService
 import com.soloist.player.service.task.TaskService
 import org.slf4j.LoggerFactory
